requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic
pymysql
pyodbc
pydantic[email]
langchain-groq
